move BitStream to new file
